requirements.txt
fastapi==0.109.0
uvicorn==0.24.0

sentence-transformers==2.2.2
faiss-cpu==1.7.4

llama-cpp-python==0.2.32

numpy==1.26.4
pandas==2.2.0
tqdm==4.66.1
python-dotenv==1.0.0

requests==2.31.0
